[redesign/refactor][ui] Improve horizontal widget responsiveness
