+++ src/ui/input/bag/BasicBagWidget.hpp
@@ -12,7 +12,6 @@ class BagTreeWidget;
 
 class QLabel;
 class QTreeWidgetItem;
-class QVBoxLayout;
 
 // Widget for basic bag tree display, without any additional target file handling
 class BasicBagWidget : public BasicInputWidget
@@ -49,7 +48,6 @@ protected slots:
 protected:
     QPointer<BagTreeWidget> m_treeWidget;
     QPointer<QLabel> m_unselectLabel;
-    QPointer<QVBoxLayout> m_controlsLayout;
 
     static constexpr int COL_CHECKBOXES = 0;
     static constexpr int COL_TOPIC_NAME = 1;
