+++ src/ui/input/bag/RecordBagWidget.cpp
@@ -106,6 +106,7 @@ RecordBagWidget::RecordBagWidget(Parameters::RecordBagParameters& parameters, QW
     advancedOptionsWidget->setLayout(advancedOptionsFormLayout);
     advancedOptionsWidget->setVisible(m_parameters.showAdvancedOptions);
 
+    m_controlsLayout->addSpacing(30);
     m_controlsLayout->addWidget(m_unselectLabel);
     m_controlsLayout->addWidget(m_treeWidget);
     m_controlsLayout->addLayout(refreshButtonLayout);
