+++ src/ui/input/helper/TopicListingInputWidget.hpp
@@ -15,7 +15,6 @@ class QFormLayout;
 class QHBoxLayout;
 class QLabel;
 class QToolButton;
-class QVBoxLayout;
 
 // A widget showing widgets to add and remove topics
 class TopicListingInputWidget : public BasicInputWidget
@@ -57,7 +56,6 @@ protected:
 
     QPointer<QFormLayout> m_formLayout;
     QPointer<QHBoxLayout> m_topicButtonLayout;
-    QPointer<QVBoxLayout> m_controlsLayout;
 
     int m_numberOfTopics = 0;
